Basic functionality
Created the NFStepper class itself with basic functionality
such as setting limits for the stepper, increasing and decreasing
it's value. We need to add the @objc attribute to access the method
runtime, because the compiler doesn't insert it automatically if the
access modifier is private. More on this topic: goo.gl/NsYD6Q